+++ templates/community-viz/package.json
@@ -1,19 +1,19 @@
 {
   "config": {
-    "gcsDevBucket": "gs://yulanlin-cv-dev/template",
+    "gcsDevBucket": "{{DEV_BUCKET}}",
     "gcsProdBucket": "{{PROD_BUCKET}}",
     "jsFile": "index.js",
     "jsonFile": "index.json",
     "cssFile": "index.css",
-    "print": "printMessage.js",
-    "manifestFile": "manifest.json"
+    "manifestFile": "manifest.json",
+    "print": "printMessage.js"
   },
   "scripts": {
     "build:dev": "node scripts/build.js --prod false",
     "build:prod": "node scripts/build.js --prod true",
     "push:dev": "node scripts/push.js --prod false",
     "push:prod": "node scripts/push.js --prod true",
-    "message": "node scripts/message.js --prod false",
+    "getMessage": "node scripts/message.js && node scripts/push.js --prod false",
     "watch": "webpack --watch",
     "start": "webpack-dev-server --open"
   },
@@ -26,7 +26,6 @@
     "mini-css-extract-plugin": "^0.5.0",
     "postcss-loader": "^3.0.0",
     "shelljs": "^0.8.3",
-    "uglifyjs-webpack-plugin": "^2.1.1",
     "webpack": "^4.26.1",
     "webpack-cli": "^3.1.2",
     "webpack-dev-server": "^3.1.10",
