+++ templates/community-viz/scripts/printMessage.js
