+++ templates/community-viz/src/index.js
@@ -1,121 +1,17 @@
 const dscc = require('@google/dscc');
 
 // import the local data
-import * as local from './localData.js';
+import * as utils from '../scripts/vizUtils.js';
+import * as viz from '../scripts/initialViz.js';
 
 // change this to 'true' for local development
 // change this to 'false' before deploying
-const LOCAL = true;
-
-const README = `
-<h2>Community Viz Local Development</h2>
-<p>
-To edit the visualization and see the changes here, edit <code>src/index.js</code>.
-</p>
-<p>
-To change the data:
-<ol>
-<li>Make sure the config in <code>src/index.json</code> is correct</li>
-<li>Run <code>npm run message</code> from the command-line </li>
-<li>Use your "dev bucket" as the component ID and deploy the visualization in Data Studio</li>
-<li>Copy the printed message + code into <code>src/localData.js</code></li>
-</ol>
-</p>
-
-<p style="margin-bottom:1cm;">
-
-</p>
-
-`;
-
-if (LOCAL) {
-  setTimeout(() => {
-    window.postMessage(local.message, '*');
-  }, 1000);
-}
+export const LOCAL = true;
 
 // write viz code here
 const drawViz = (data) => {
-  // append a div to the DOM
-  if (document.querySelector('div')) {
-    var oldDiv = document.querySelector('div');
-    oldDiv.parentNode.removeChild(oldDiv);
-  }
-
-  if (document.querySelector('canvas')) {
-    var oldCanv = document.querySelector('canvas');
-    oldCanv.parentNode.removeChild(oldDiv);
-  }
-
-  // append the data to a div
-  let readmeDiv = document.createElement('div');
-  readmeDiv.innerHTML = README;
-  document.body.appendChild(readmeDiv);
-
-  // everything below is verbatim from the codelab
-  // create and add the canvas
-  var canvasElement = document.createElement('canvas');
-  var ctx = canvasElement.getContext('2d');
-  canvasElement.id = 'myViz';
-  document.body.appendChild(canvasElement);
-
-  var rowData = data.tables.DEFAULT;
-  var ctx = canvasElement.getContext('2d');
-
-  // clear the canvas.
-  ctx.clearRect(0, 0, canvasElement.width, canvasElement.height);
-
-  // set the canvas width and height
-  ctx.canvas.width = 300;
-  ctx.canvas.height = 200;
-
-  // scale the bar width and max bar height to the canvas
-  var barWidth = ctx.canvas.width / (rowData.length * 2);
-  var maxBarHeight = ctx.canvas.height - 20;
-
-  // vertical offset for bar text
-  var textYOffset = 20;
-
-  ctx.fillStyle = '#000000';
-
-  // obtain the maximum bar metric value for scaling purposes
-  var metricMax = 0;
-  rowData.forEach(function(row) {
-    metricMax = Math.max(metricMax, row['metricID'][0]);
-  });
-
-  // draw bars
-  // add dimension labels below bars
-  // 'barDimension' and 'barMetric' come from the id defined in myViz.json
-  rowData.forEach(function(row, i) {
-    // calculates the height of the bar using the row value, maximum bar
-    // height, and the maximum metric value calculated earlier
-    var barHeight = Math.round(
-      -1 * ((row['metricID'][0] * maxBarHeight) / metricMax)
-    );
-
-    // calculates the x coordinate of the bar based on the width of the convas
-    // and the width of the bar
-    var barX = (ctx.canvas.width / rowData.length) * i + barWidth / 2;
-
-    ctx.fillRect(barX, maxBarHeight, barWidth, barHeight);
-
-    var barText = row['dimID'][0];
-    var textX = barX + barWidth / 4;
-    var textY = maxBarHeight + textYOffset;
-
-    ctx.fillText(barText, textX, textY);
-  });
+  viz.firstViz(data);
 };
 
 // logic to handle local vs. Data Studio deployment
-if (LOCAL) {
-  const handleMessage = (message) => {
-    if (!message.data.type) {
-      drawViz(message.data);
-    }
-  };
-  window.addEventListener('message', handleMessage);
-} else {
-  dscc.subscribeToData(drawViz, {transform: dscc.objectTransform});
-}
+utils.renderViz(LOCAL, drawViz);
